clipa_jax/requirements-tpu.txt
absl-py==1.4.0
chex==0.1.5
clu==0.0.8
einops==0.6.0
flax==0.6.3
git+https://github.com/google/flaxformer

jax[tpu]==0.4.1
--find-links https://storage.googleapis.com/jax-releases/libtpu_releases.html

ml-collections==0.1.1
numpy>=1.21.1
pandas==1.5.2
tensorflow==2.11.0
tensorflow-datasets==4.8.2
tensorflow-probability==0.19.0
tensorflow-text==2.11.0
wandb==0.13.7
overrides==7.3.1
tensorflow_addons==0.19.0